src/main/java/com/training/spring/bigcorp/service/CaptorServiceImpl.java
package com.training.spring.bigcorp.service;

import com.training.spring.bigcorp.model.Captor;
import com.training.spring.bigcorp.model.PowerSource;
import com.training.spring.bigcorp.service.measure.MeasureService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Service;

import java.util.HashSet;
import java.util.Set;

@Service
public class CaptorServiceImpl implements CaptorService {

    private MeasureService measureService;

    @Autowired
    public CaptorServiceImpl(@Qualifier("FixedMeasureService") MeasureService measureService) {
        this.measureService = measureService;
    }

    @Override
    public Set<Captor> findBySite(String siteId) {
        Set<Captor> captors = new HashSet<>();
        if (siteId == null) {
            return captors;
        }
        captors.add(new Captor("Capteur A", PowerSource.FIXED));
        return captors;
    }

    public MeasureService getMeasureService() {
        return measureService;
    }
}
